# applications/service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Doppelkopf Service API
  description: API for the Doppelkopf service
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: GameRead
    description: Operations related to reading game data
  - name: GameWrite
    description: Operations related to creating and updating games
paths:
  /games:
    post:
      summary: Create a new game
      operationId: createGame
      tags:
        - GameWrite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameCreatedResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{gameId}:
    get:
      summary: Get game by ID
      operationId: getGame
      tags:
        - GameRead
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier of the game
      responses:
        '200':
          description: Game retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{gameId}/play:
    post:
      summary: Play a card on the game.
      operationId: playCard
      tags:
        - GameWrite
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier of the game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayCardRequest'
      responses:
        '200':
          description: Card played successfully on game.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{gameId}/contracts/propose:
    post:
      summary: Propose a contract, i.e. a different kind of game is to be played
      operationId: proposeContract
      tags:
        - GameWrite
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier of the game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposeContractRequest'
      responses:
        '200':
          description: Contract proposed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}/contracts/accept-armut:
    post:
      summary: Accept an already proposed Armut.
      operationId: acceptArmut
      tags:
        - GameWrite
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier of the game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptArmutRequest'
      responses:
        '200':
          description: Armut accepted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}/contracts/pass:
    post:
      summary: Pass on proposing contracts. If all players pass, a normal game is played.
      operationId: passOnContractPropose
      tags:
        - GameWrite
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier of the game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PassOnContractsRequest'
      responses:
        '200':
          description: Passing on contracts succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}/announcement:
    post:
      summary: Make an announcement, e.g. Re, Contra.
      operationId: announce
      tags:
        - GameWrite
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier of the game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnouncementRequest'
      responses:
        '200':
          description: Announcement successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    GameResponse:
      type: object
      required:
        - id
        - name
        - players
      properties:
        id:
          type: string
          description: The unique identifier of the game
        currentPlayer:
          $ref: '#/components/schemas/PlayerId'
        bout:
          $ref: '#/components/schemas/Bout'
        playerHands:
          type: array
          items:
            $ref: '#/components/schemas/PlayerHand'

    Bout:
      type: object
      required:
        - cards
      properties:
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'

    GameCreatedResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: The unique identifier of the created game

    PlayerId:
      type: string
      description: The unique identifier of a player

    PlayerHand:
      type: object
      required:
        - playerId
        - cards
      properties:
        playerId:
          $ref: '#/components/schemas/PlayerId'
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'

    Card:
      type: object
      required:
        - suit
        - rank
      properties:
        suit:
          type: string
          description: The suit of the card (e.g., Hearts, Diamonds, Clubs, Spades)
        rank:
          type: string
          description: The rank of the card (e.g., 2, 3, 4, 5, 6, 7, 8, 9, 10, J, Q, K, A)

    CreateGameRequest:
      type: object
      required:
        - playerIds
      properties:
        playerIds:
          type: array
          items:
            $ref: '#/components/schemas/PlayerId'
        withNines:
          type: boolean
          default: false

    PlayCardRequest:
      type: object
      required:
        - playerId
        - card
      properties:
        playerId:
          $ref: '#/components/schemas/PlayerId'
        card:
          $ref: '#/components/schemas/Card'

    AnnouncementRequest:
      type: object
      required:
        - playerId
        - announcement
      properties:
        playerId:
          $ref: '#/components/schemas/PlayerId'
        announcement:
          type: string
          enum:
            - RE
            - CONTRA
            - UNTER_90
            - UNTER_60
            - UNTER_30
            - SCHWARZ

    ProposeContractRequest:
      type: object
      required:
        - playerId
        - proposition
      properties:
        playerId:
          $ref: '#/components/schemas/PlayerId'
        proposition:
          type: string
          enum:
            - HOCHZEIT
            - ARMUT
            - OBER_SOLO
            - UNTER_SOLO
            - FLEISCHLOS
            - TRUMPF_SOLO
            - HEARTS_SOLO
            - SPADES_SOLO
            - CLUBS_SOLO
        cards:
          type: string
          description: Only present on ARMUT.

    AcceptArmutRequest:
      type: object
      required:
        - playerId
        - returnedTrumpCards
      properties:
        playerId:
          $ref: '#/components/schemas/PlayerId'
        returnedTrumpCards:
          type: array
          items:
            $ref: '#/components/schemas/Card'

    PassOnContractsRequest:
      type: object
      required:
        - playerId
      properties:
        playerId:
          type: string

    ErrorResponse:
      type: object
      required:
        - status
        - error
        - message
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string